Move all logic from component to new, singleton service
